backend/pyproject.toml
[tool.poetry]
name = "pronielts-backend"
version = "1.0.0"
description = "IELTS Pronunciation Assessment Platform - Backend API"
authors = ["Mesaque Junior <mesaquejunior@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
azure-cognitiveservices-speech = "^1.34.0"
azure-storage-blob = "^12.19.0"
azure-identity = "^1.15.0"
cryptography = "^42.0.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.26.0"
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.1.1"
pylint = "^3.0.3"
faker = "^22.0.0"
ipython = "^8.20.0"
isort = "^5.13.2"
ruff = "^0.8.5"
mypy = "^1.13.0"
deptry = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | alembic/versions
)/
'''

[tool.pylint.messages_control]
max-line-length = 100
disable = ["C0111", "C0103", "R0903"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["app"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "alembic/versions",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
    "B904", # raise from exception (can be enabled later for better error handling)
    "C416", # unnecessary dict comprehension
]

[tool.mypy]
python_version = "3.13"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = false
warn_redundant_casts = false
warn_unused_ignores = false
warn_no_return = false
follow_imports = "normal"
ignore_missing_imports = true
exclude = ["alembic/versions/", "tests/"]
# Start with lenient settings, gradually increase strictness

[tool.deptry]
exclude = ["tests", "alembic"]
# Ignore specific dependencies that will be used in future features or are loaded dynamically
per_rule_ignores = { DEP002 = [
    "alembic",  # Database migrations (used via CLI)
    "psycopg2-binary",  # PostgreSQL driver (loaded by SQLAlchemy)
    "python-multipart",  # File upload support (used by FastAPI)
    "python-jose",  # JWT tokens (future auth feature)
    "passlib",  # Password hashing (future auth feature)
    "httpx",  # Async HTTP client (future feature)
    "email-validator",  # Email validation (future feature)
]}
